fix: fixed test regressions + refactoring
