+++ lib/glific/repo_replica.ex
@@ -3,11 +3,19 @@ defmodule Glific.RepoReplica do
   A read-replica repository that maps to an underlying data store, controlled by the Postgres adapter.
   """
 
+  # In tests replica uses primary's connection pool
+  dynamic_repo =
+    if Application.compile_env!(:glific, :environment) == :test do
+      Glific.Repo
+    else
+      __MODULE__
+    end
+
   use Ecto.Repo,
     otp_app: :glific,
-    adapter: Ecto.Adapters.Postgres
+    adapter: Ecto.Adapters.Postgres,
+    default_dynamic_repo: dynamic_repo,
+    read_only: true
 
-  # TODO: if we are adding `read_only: true`, then RepoHelpers can't have delete fn, so we need
-  # to figure out a way to not add the function in that case, for now its fine. 
   use Glific.RepoHelpers
 end
